Move Generate from usage tests to new framework
