+++ src/VisualStudio/IntegrationTest/New.IntegrationTests/CSharp/CSharpGenerateFromUsage.cs
@@ -4,6 +4,7 @@
 
 #nullable disable
 
+using System.Threading.Tasks;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.Test.Utilities;
 using Microsoft.VisualStudio.IntegrationTest.Utilities;
@@ -13,29 +14,28 @@ using Xunit.Abstractions;
 
 namespace Roslyn.VisualStudio.IntegrationTests.CSharp
 {
-    [Collection(nameof(SharedIntegrationHostFixture))]
     public class CSharpGenerateFromUsage : AbstractEditorTest
     {
         protected override string LanguageName => LanguageNames.CSharp;
 
-        public CSharpGenerateFromUsage(VisualStudioInstanceFactory instanceFactory)
-            : base(instanceFactory, nameof(CSharpGenerateFromUsage))
+        public CSharpGenerateFromUsage()
+            : base(nameof(CSharpGenerateFromUsage))
         {
         }
 
-        [WpfFact, Trait(Traits.Feature, Traits.Features.CodeActionsGenerateLocal)]
-        public void GenerateLocal()
+        [IdeFact, Trait(Traits.Feature, Traits.Features.CodeActionsGenerateLocal)]
+        public async Task GenerateLocal()
         {
-            SetUpEditor(
+            await SetUpEditorAsync(
 @"class Program
 {
     static void Main(string[] args)
     {
         string s = $$xyz;
     }
-}");
-            VisualStudio.Editor.Verify.CodeAction("Generate local 'xyz'", applyFix: true);
-            VisualStudio.Editor.Verify.TextContains(
+}", HangMitigatingCancellationToken);
+            await TestServices.EditorVerifier.CodeActionAsync("Generate local 'xyz'", applyFix: true, cancellationToken: HangMitigatingCancellationToken);
+            await TestServices.EditorVerifier.TextContainsAsync(
 @"class Program
 {
     static void Main(string[] args)
@@ -43,7 +43,7 @@ namespace Roslyn.VisualStudio.IntegrationTests.CSharp
         string xyz = null;
         string s = xyz;
     }
-}");
+}", cancellationToken: HangMitigatingCancellationToken);
         }
     }
 }
